Fixes how Ohm's toAST handles lexical rules with semantic actions somewhere under the tree
The built-in toAST semantic only handles lexical rules iff a semantic action exists for *every* node within the tree. That's a lot of work. This keeps some of the implicit AST definition for lexical rules by letting semantic actions to be omitted for rules that only have a single children (alternatives/etc).